update demo code

+++ src/networkimage/src/NINetworkImageView.m
@@ -357,11 +357,13 @@
          [self _didFailToLoadWithError:error];
       }];
 
-      [requestOperation setDownloadProgressBlock:^(NSUInteger bytesRead, NSInteger totalBytesRead, NSInteger totalBytesExpectedToRead) {
-          if ([self.delegate respondsToSelector:@selector(networkImageView:readBytes:totalBytes:)]) {
-              [self.delegate networkImageView:self readBytes:totalBytesRead totalBytes:totalBytesExpectedToRead];
-          }
-      }];
+        [requestOperation setDownloadProgressBlock:^(NSUInteger bytesRead, long long totalBytesRead, long long totalBytesExpectedToRead) {
+            
+            if ([self.delegate respondsToSelector:@selector(networkImageView:readBytes:totalBytes:)]) {
+                [self.delegate networkImageView:self readBytes:totalBytesRead totalBytes:totalBytesExpectedToRead];
+            }
+            
+        }];
 
       self.operation = requestOperation;
